// src/types/types.ts
/**
 * 型ファイル
 */
// 表示パターン型
export type CircleInfo = {
  number: number;
  top: string;
  left: string;
  width: string;
  height: string;
  fontSize: string;
  backgroundColor: string;
};
// ゲーム結果型
export type GameResult = {
  correct: number;
  wrong: number;
}